Added branch support for local repositories
Need to add a "git-jira-hook.branches" git config
pointing to a list of branch names.
If this list exists, then the hook will only be enabled
if the current commit is on one of the named branches.

Else, hook will stay disabled.

By default, if config is empty, then hook will stay enabled
for all branches.